requirements.txt
tensorflow
opencv-python-headless
numpy
pillow
requests